+++ src/Lesson_2/ArraySum/MyArraySizeException.java
@@ -1,4 +1,4 @@
-package Lesson_2;
+package Lesson_2.ArraySum;
 
 public class MyArraySizeException extends Exception {
